refactor(core/utils): transfer normalizePadding() to core/markdown
