+++ src/core/webidl.js
@@ -5,9 +5,9 @@
 //  - It could be useful to report parsed IDL items as events
 //  - don't use generated content in the CSS!
 import * as webidl2 from "webidl2";
-import { flatten, normalizePadding } from "./utils.js";
 import css from "text!../../assets/webidl.css";
 import { findDfn } from "./dfn-finder.js";
+import { flatten } from "./utils.js";
 import hyperHTML from "hyperhtml";
 import { pub } from "./pubsubhub.js";
 import { registerDefinition } from "./dfn-map.js";
@@ -296,7 +296,7 @@ export function run() {
         "error",
         `Failed to parse WebIDL: ${e.message}.
         <details>
-        <pre>${normalizePadding(idlElement.textContent)}\n ${e}</pre>
+        <pre>${idlElement.textContent}\n ${e}</pre>
         </details>`
       );
       // Skip this <pre> and move on to the next one.
